# openapi.yaml
openapi: 3.1.0
info:
  title: REASON Docs Management API
  description: |
    A comprehensive API for the REASON note-taking application. This API provides endpoints for managing documents, teams, organizations, and user settings.

    ## Features
    - 📝 Document management with hierarchical structure
    - 👥 Team and organization collaboration
    - 🔍 Full-text search capabilities
    - 🔄 Bulk operations and syncing
    - 🎨 Theme and settings management
servers:
  - url: https://editor.qwksearch.com/api
    description: Production server
...
paths:
...
  /api/documents:
    get:
      tags:
        - Documents
      summary: Get all documents
      description: Retrieve a list of all documents in the system
      operationId: getAllDocuments
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/DocumentListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/ErrorResponse'

    post:
      tags:
        - Documents
      summary: Create a new document
      description: Create a new document with optional parent
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/editor/schemas/CreateDocumentRequest'
            examples:
              root-document:
                summary: Create root document
                value:
                  title: "My New Note"
                  content: "<p>This is my note content</p>"
              child-document:
                summary: Create child document
                value:
                  title: "Child Note"
                  content: "<p>This is a child note</p>"
                  parentId: "123"
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/DocumentResponse'
        '400':
          description: Bad request (missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/ErrorResponse'
...
  /api/google-docs/export:
    post:
      tags:
        - Google Docs
      summary: Export document to Google Docs
      description: Export a REASON document to Google Docs
      operationId: exportToGoogleDocs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                documentId:
                  type: string
                  description: ID of the document to export
                accessToken:
                  type: string
                  description: Google OAuth access token
                refreshToken:
                  type: string
                  description: Google OAuth refresh token
              required:
                - documentId
                - accessToken
      responses:
        '200':
          description: Successfully exported
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      documentId:
                        type: string
                      webViewLink:
                        type: string
                      webContentLink:
                        type: string
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/editor/schemas/ErrorResponse'
  /api/google-docs/import:
    post:
      tags:
        - Google Docs
      summary: Import document from Google Docs
      description: Import a Google Docs document into REASON
      operationId: importFromGoogleDocs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                googleDocId:
                  type: string
                  description: Google Docs document ID
                accessToken:
                  type: string
                  description: Google OAuth access token
                refreshToken:
                  type: string
                  description: Google OAuth refresh token
              required:
                - googleDocId
                - accessToken
      responses:
        '200':
          description: Successfully imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/editor/schemas/Document'
...